tasks/research_assistant/task5/assets/input/codebase/requirements.txt
torch==2.1.0
transformers==4.35.0
Pillow>=9.0.0
numpy>=1.24.0
scipy
scikit-learn>=1.3.0
nltk>=3.8
pycocoevalcap>=1.2
pyyaml>=6.0
tqdm>=4.65.0
